+++ test/test_pooling.py
@@ -169,7 +169,7 @@ class _TestPoolingBase(IntegrationTest):
             **kwargs):
         # Start the pool with the correct ssl options.
         pool_options = client_context.client._topology_settings.pool_options
-        kwargs['ssl_context'] = pool_options.ssl_context
+        kwargs['ssl_context'] = pool_options._ssl_context
         kwargs['tls_allow_invalid_hostnames'] = pool_options.tls_allow_invalid_hostnames
         kwargs['server_api'] = pool_options.server_api
         pool = Pool(pair, PoolOptions(*args, **kwargs))
@@ -187,7 +187,7 @@ class TestPooling(_TestPoolingBase):
             ValueError, MongoClient, host=host, port=port, maxPoolSize='foo')
 
         c = MongoClient(host=host, port=port, maxPoolSize=100, connect=False)
-        self.assertEqual(c.max_pool_size, 100)
+        self.assertEqual(c.options.pool_options.max_pool_size, 100)
 
     def test_no_disconnect(self):
         run_cases(self.c, [NonUnique, Unique, InsertOneAndFind])